study: mybatis ìCRUD 실습 id type ë³Long 으로 변경

+++ java/exam_mybatis/src/main/java/codingon/exam_mybatis/controller/BoardController.java
@@ -30,7 +30,7 @@ public class BoardController {
 
     // 특정 ID 의 게시글 정보 반환
     @GetMapping("/{id}")
-    public BoardDTO getBoard(@PathVariable int id) { return boardService.getBoardById(id); }
+    public BoardDTO getBoard(@PathVariable Long id) { return boardService.getBoardById(id); }
 
     // 새 게시글 생성하고 생성된 게시글 정보 반환
     @PostMapping
@@ -41,7 +41,7 @@ public class BoardController {
 
     // 특정 ID 의 게시글 정보를 업데이트하고 업데이트된 정보를 반환
     @PutMapping("/{id}")
-    public BoardDTO updateBoard(@PathVariable int id, @RequestBody BoardDTO boardDTO) {
+    public BoardDTO updateBoard(@PathVariable Long id, @RequestBody BoardDTO boardDTO) {
         boardDTO.setId(id);
         boardService.updateBoard(boardDTO);
         return boardDTO;
@@ -49,5 +49,5 @@ public class BoardController {
 
     // 특정 ID 의 게시글 삭제
     @DeleteMapping("/{id}")
-    public void deleteBoard(@PathVariable int id) { boardService.deleteBoard(id); }
+    public void deleteBoard(@PathVariable Long id) { boardService.deleteBoard(id); }
 }
